+++ mayan/apps/acls/tests/test_links.py
@@ -1,9 +1,8 @@
 from __future__ import unicode_literals
 
-from django.contrib.contenttypes.models import ContentType
 from django.urls import reverse
 
-from mayan.apps.documents.tests import GenericDocumentViewTestCase
+from mayan.apps.common.tests import GenericViewTestCase
 
 from ..links import (
     link_acl_delete, link_acl_list, link_acl_create, link_acl_permissions
@@ -13,40 +12,34 @@ from ..permissions import permission_acl_edit, permission_acl_view
 from .mixins import ACLTestMixin
 
 
-class ACLsLinksTestCase(ACLTestMixin, GenericDocumentViewTestCase):
+class ACLsLinksTestCase(ACLTestMixin, GenericViewTestCase):
     def setUp(self):
         super(ACLsLinksTestCase, self).setUp()
-        self.test_object = self.test_document
+        self._setup_test_object()
 
-    def test_document_acl_create_link(self):
+    def test_object_acl_create_link(self):
         self.grant_access(
-            obj=self.test_document, permission=permission_acl_edit
+            obj=self.test_object, permission=permission_acl_edit
         )
 
-        self.add_test_view(test_object=self.document)
+        self.add_test_view(test_object=self.test_object)
         context = self.get_test_view()
         resolved_link = link_acl_create.resolve(context=context)
 
         self.assertNotEqual(resolved_link, None)
 
-        content_type = ContentType.objects.get_for_model(self.test_document)
-        kwargs = {
-            'app_label': content_type.app_label,
-            'model': content_type.model,
-            'object_id': self.document.pk
-        }
-
         self.assertEqual(
             resolved_link.url, reverse(
-                viewname=link_acl_create.view, kwargs=kwargs
+                viewname=link_acl_create.view,
+                kwargs=self.test_content_object_view_kwargs
             )
         )
 
-    def test_document_acl_delete_link(self):
+    def test_object_acl_delete_link(self):
         self._create_test_acl()
 
         self.grant_access(
-            obj=self.test_document, permission=permission_acl_edit
+            obj=self.test_object, permission=permission_acl_edit
         )
 
         self.add_test_view(test_object=self.test_acl)
@@ -61,11 +54,11 @@ class ACLsLinksTestCase(ACLTestMixin, GenericDocumentViewTestCase):
             )
         )
 
-    def test_document_acl_edit_link(self):
+    def test_object_acl_edit_link(self):
         self._create_test_acl()
 
         self.grant_access(
-            obj=self.test_document, permission=permission_acl_edit
+            obj=self.test_object, permission=permission_acl_edit
         )
 
         self.add_test_view(test_object=self.test_acl)
@@ -81,25 +74,19 @@ class ACLsLinksTestCase(ACLTestMixin, GenericDocumentViewTestCase):
             )
         )
 
-    def test_document_acl_list_link(self):
+    def test_object_acl_list_link(self):
         self.grant_access(
-            obj=self.test_document, permission=permission_acl_view
+            obj=self.test_object, permission=permission_acl_view
         )
 
-        self.add_test_view(test_object=self.test_document)
+        self.add_test_view(test_object=self.test_object)
         context = self.get_test_view()
         resolved_link = link_acl_list.resolve(context=context)
         self.assertNotEqual(resolved_link, None)
 
-        content_type = ContentType.objects.get_for_model(self.test_document)
-        kwargs = {
-            'app_label': content_type.app_label,
-            'model': content_type.model,
-            'object_id': self.test_document.pk
-        }
-
         self.assertEqual(
             resolved_link.url, reverse(
-                viewname=link_acl_list.view, kwargs=kwargs
+                viewname=link_acl_list.view,
+                kwargs=self.test_content_object_view_kwargs
             )
         )